Practic 6

+++ tgbot/misc/throttling.py
@@ -1,4 +1,4 @@
-def rate_limit(limit: int, key=None):
+def rate_limit(limit: float, key=None):
     def decorator(func):
         setattr(func, "throttling_rate_limit", limit)
         if key:
@@ -6,4 +6,3 @@ def rate_limit(limit: int, key=None):
         return func
 
     return decorator
-
